airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyspark==3.5.0
apache-airflow-providers-apache-spark
apache-airflow-providers-amazon
python-dotenv